third_party/blink/web_tests/platform/generic/virtual/document-transition-wide-gamut/document-transition/capture-callback-exception-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: foo is not defined
